Set up aws (#11)
## Description of the PR
We want to have the app deployed so that others can follow along with
the progress and test.

## Solution Description
We'll start with using aws since there's a decent free tier. 
The flask app can run on a lambda for now since those requests are
relatively quick to process
We can use dynamo db for storage since we get 25gb free.

The app sets up using a sam template for ease of creation and removal of
resources.
The structure of the app was reconfigured so that the app loads both
locally and in aws.

## Tests
No testable changes

## Manual Testing
Yes deployed both locally and on aws and verified that the app is
loading as expected.