Get Geocoding Api Working (#12)
## JIRA Task Title And Link
[Get Geocoding api
working](https://ideapartners.atlassian.net/browse/JOUR-1)

## Description of the PR
User input addresses need to be geocoded so that we can validate them
and ultimately render them on a map.

## Solution Description
We use mapboxes api which accepts a search string and returns 5
locations with metadata.
We debounce the requests to have 2 seconds of time between them and then
render results as suggestions.
Suggestions appear below the address and when a user makes a selection
we store that object along with the it's coordinates.
If no locations are returned from mapbox, we render an error message. 

## Tests
Nothing added yet.

## Manual Testing
Tested manually in localhost.