JOUR-5 - Add custom icons (#15)
## JIRA Task Title And Link
[Add custom icons](https://ideapartners.atlassian.net/browse/JOUR-5)

## Description of the PR
We want to be able to have custom icons for our points on the map.

## Solution Description
Here we use the openlayers options to set a custom icon as our point and
also update our vite config to migrate the images to the proper
directory for flask to recognize where they are.

## Tests
Not yet. Still experimenting with just figuring out the limitations and
capabilities of the map.

## Manual Testing
Yes, tested locally.