JOUR-6 - Added dropdown to choose map tile layer (#16)
## JIRA Task Title And Link
[choose filters for the
map](https://ideapartners.atlassian.net/browse/JOUR-6)

## Description of the PR
We want users to be able to choose a map style that they like

## Solution Description
- We use Stadia maps for now and take a few of the free styles. 
- We add a dropdown where users can make a selection
- We then update the maps layer source with the dropdown selection.

## Tests
Not yet. Still just learning the possibilities of the map

## Manual Testing
Tested locally.

## Other comments
There could be other options that we can use including osm layers and
Nextzen vector layers.
With the vector layers we might even be able to let users make more
serious decisions like color of water, roads, thicknesses, etc. But that
will need to be explored.