JOUR-10 - Tidying up (minor refactors, comments, reorganization) (#17)
## JIRA Task Title And Link
[Clean up](https://ideapartners.atlassian.net/browse/JOUR-10)

## Description of the PR
We’ve moved fast the past few sprints, so we want to go through the code
and clean it up a bit, add the docstrings, do minor refactors, etc.

Just to make sure that we don’t have too much refactoring to do at a
later time.

As part of this, we should take a look and verify if it’s time to start
adding any unit tests

## Solution Description
- We added docstrings and comments in various files to make it easier to
understand what happened.
- We reorganized the order of the javascript call so that all functions
were towards the bottom. This makes it easier to see what code is
actually being called and run.
- We factored out logic in the submit handler to use smaller named
functions that encapsulate logic.
- We also get rid of the global variable validatedLocations, and instead
store the chosen location on the address input itself using the
data-coordinates property.

## Tests
This code is still exploratory and almost none of it may stay, so no
tests are added at this time.

## Manual Testing
The app was tested with multiple locations after the refactor to ensure
that it still works as intended.