JOUR-11 - Points are now rendered along with the line animation. (#18)
## JIRA Task Title And Link
[Plot points with
delay](https://ideapartners.atlassian.net/browse/JOUR-11)

## Description of the PR
We want points to be plotted also as the line moves along and not right
away.

## Solution Description
We don't render the points immediately and instead render them with a
3second delay which matches the animation of the lines.

Also realized that we checked in a bug with the last pr where we started
storing the coordinates on the data-coordinates object of the element
itself. This caused it to be cast to a string and thus error out in the
functions that expected an array of coordinates. Now we are using
JSON.stringify to keep it's type.

## Tests
Still exploring

## Manual Testing
Yes tested manually to verify that it is working as expected.