(JOUR-16, JOUR-18) Zoom and music (#24)
## JIRA Task Title And Link
[Add Music](https://ideapartners.atlassian.net/browse/JOUR-18)
[Smooth out Transitions](https://ideapartners.atlassian.net/browse/JOUR-16)

## Description of the PR
The map was a bit janky with it's loading as it moved from location to
location. We want it to be a bit smoother so that we have an easier time
capturing the animation.

We also wanted to add some fun to the video to add some music

## Solution Description
- Add an audio file that starts playing with the animation and ends a
little bit after. The audio loops.
- Add additional zoom levels based on distances. Previously we had 5 and
10, now we have from 4-12.
- We changed some settings around animation in the hopes of making it
smoother.
- We added some zIndexes to try and get the images to persist even as
the zoom changed. This wasn't successful, but the code was left in.
- We changed the script file to copy all of the media files from the
client to the functions folder.
- Since we now have a custom domain, we are able to use the simple
/static/... to load our media files. So we removed the string
manipulation that added url_for from the build script. We might add this
back.

## Tests
No, all still exploratory

## Manual Testing
Yes tested manually in the browser.