JOUR-28 - map functions moved to map.js and random_suggestions improved (#26)
## JIRA Task Title And Link
[Tidying up](https://ideapartners.atlassian.net/browse/JOUR-28)

## Description of the PR
After the last few sprints the code has gotten a bit messy. It's getting
hard to add functions to the javascript now since there is so much code
in there. We should split it up.

Additionally, it seems that our testing function for getting random
latitude/longitude was not returning the kind of data that helped us
actually test the map, so we should fix it to return valid coordinates
of cities and places.

## Solution Description
- `main.js` was refactored so that all of the map functions moved to
`map.js` and imported into `main.js`
- Only DOM related functions are still in `main.js` now. 
- actual location suggestions from mapbox were recorded and captured in
an array.
- we create a function `get_n_random_suggestions` that returns `n`
suggestions from the array.


## Tests
Still in exploratory code.

## Manual Testing
Yes, verified that the application is still running after refactor and
with random suggestions.