JOUR-20 - photos are stored in filemap and accessible on submit (#27)
## JIRA Task Title And Link
[Render photos during
pauses](https://ideapartners.atlassian.net/browse/JOUR-20)

## Description of the PR
We want to show photos as we are pausing at each individual city.

## Solution Description
- When we create the thumbnail, we now also store the loaded image in
the filemap object. When we submit, we pull the image from that object
to be able to pass it into the map functions. There is a risk that
someone would then decide not to use that image and we would still be
storing it in memory. But I think if we are also removing images after
we have loaded things into the dom and done the animation, then it might
not be such a bad thing.
- Images are added to the map in a cross pattern around the location
(currently not well situated). The images are rendered as polaroids and
we are just using the point coordinates as the base and specifying an
offset. This means that the images are staying exactly where they are
across zoom levels, just like we can still see the icon of the place
regardless of zoom level. Currently the images are not looking too great
and don't have any animations or fade, and they are not being removed,
cluttering the map.

## Tests
Still exploring

## Manual Testing
Verified locally.