JOUR-22 - 18 * 24 png of map generated on button click (#29)
## JIRA Task Title And Link
Capture map as image

## Description of the PR
We want to be able to save the map as a large scale image so that we can
print it on a poster

## Solution Description
We redraw the map in an offscreen canvas where we set the dimensions
We add a button to download the png of the current rendered map.
We save the offscreen canvas to a png file and download it. 

## Tests
Still exploring

## Manual Testing
Yes, tested locally.